% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_test.R
\docType{data}
\name{long_test}
\alias{long_test}
\title{A sample of long form of testing data}
\format{data.frame}
\usage{
data(long_test)
}
\description{
This contains the long form of the simulated testing data.
}
\examples{
data(long_test)
}
\keyword{datasets}
