% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_train.R
\docType{data}
\name{long_train}
\alias{long_train}
\title{A sample of long form of training data}
\format{data.frame}
\usage{
data(long_train)
}
\description{
This contains the long form of the simulated training data.
}
\examples{
data(long_train)
}
\keyword{datasets}
