% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictSurvProb.R
\name{predictor_loader}
\alias{predictor_loader}
\title{predictor_loader}
\usage{
predictor_loader()
}
\value{
None
}
\description{
An intermediate function for loading the necessary function into .GlobalEnv
}
\examples{
# must run the following code before calling funest_pred()
predictSurvProb.ranger = predictor_loader()
}
