% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_test.R
\docType{data}
\name{surv_test}
\alias{surv_test}
\title{A sample of short form of testing data}
\format{data.frame}
\usage{
data(surv_test)
}
\description{
This contains the short form of the simulated testing data.
}
\examples{
data(surv_test)
}
\keyword{datasets}
