% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_train.R
\docType{data}
\name{surv_train}
\alias{surv_train}
\title{A sample of short form of training data}
\format{data.frame}
\usage{
data(surv_train)
}
\description{
This contains the short form of the simulated training data.
}
\examples{
data(surv_train)
}
\keyword{datasets}
